rename memoryData to state
